product-research/flipsnack-alternative/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
pypdf2==3.0.1
pillow==10.1.0
reportlab==4.0.7
python-multipart==0.0.6
minio==7.2.0
boto3==1.34.0
sentry-sdk==1.35.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
aiofiles==23.2.1